Refactor Trigger classes
Type: trivial
