+++ pgtrigger/triggers/__init__.py
@@ -1,0 +1,7 @@
+from .fsm import FSM
+from .protect import Protect
+from .softdelete import SoftDelete
+from .trigger import Trigger
+from .updatesearchvector import UpdateSearchVector
+
+__all__ = [FSM, Protect, SoftDelete, Trigger, UpdateSearchVector]
